Split out menu (#10)
* split out menu

* lint
